requirements.txt
# Interface utilisateur
streamlit>=1.28.0

# Visualisations
plotly>=5.15.0

# Manipulation de données
pandas>=1.5.0
numpy>=1.24.0

# Communication API
requests>=2.28.0

# Configuration environnement
python-dotenv>=1.0.0